working on creating operation procedures

+++ db/DDL/tables/latest_balances.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS latest_balances(
     accountNumber INTEGER REFERENCES account(accountNumber),
-    amount BIGINT
+    amount NUMERIC(18, 0)
 );
